+++ R/capHistSum.R
@@ -5,11 +5,16 @@
 #' @details This function requires the capture history data file to be in the \dQuote{individual} fish format.  See \code{\link{capHistConvert}} for a description of this (and other) formats and for methods to convert from other formats to the \dQuote{individual} fish format.  In addition, this function requires only the capture history portion of the data file.  Thus, if \code{df} contains columns with non-capture history information (e.g., fish ID, length, location, etc.) then use \code{cols2use=} to identify which columns contain only the capture history information.  Columns to use can be identified by listing the column numbers (e.g., columns 2 through 7 could be included with \code{cols2use=2:7}).  In many instances it may be easier to identify columns to \emph{exclude} which can be done by preceding the column number by a negative sign (e.g., columns 1 through 3 are excluded with \code{cols2use=-(1:3)}).
 #' 
 #' The object returned from this function can be used directly in \code{\link{mrClosed}} and \code{\link{mrOpen}}.  See examples of this functionality on the help pages for those functions.
+#' 
+#' The \code{plot} function can be used to construct the two diagnostic plots described by Baillargeon and Rivest (2007).  The \code{what="f"} plot will plot the log of the number of fish seen i times divided by \code{choose(t,i)} against i.  The \code{what="u"} plot will plot the log of the number of fish seen for the first time on event i against i.  Baillargeon and Rivest (2007) provide a table that can be used to diagnosed types of heterogeneities in capture probabilities from these plots.
 #'
 #' @note This function assumes that all unmarked captured fish are marked and returned to the population (i.e., no losses at the time of marking are allowed).
 #' 
 #' @param df A data.frame that contains the capture histories (and, perhaps, other information) in \dQuote{individual} fish format.  See details.
 #' @param cols2use A numeric vector of columns that contain the capture histories.  See details.
+#' @param x An object from \code{capHistSum}.
+#' @param what A string that indicates what type of diagnostic plot to construct with \code{plot}.  See details.
+#' @param \dots Optional arguments to send to \code{plot}.
 #'
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}
 #'
@@ -22,27 +27,37 @@
 #' If the capture history data file represents more than two samples, then a list with the following five components is returned
 #'  \itemize{
 #'    \item \code{caphist} A vector summarizing the frequency of fish with each capture history.
-#'    \item \code{sum} A data frame that contains the the number of captured fish in the ith sample (\code{n}), the number of recaptured (i.e. previously marked) fish in the ith sample (\code{m}), the number of marked fish returned to the population following the ith sample (\code{R}; this will equal \code{n} as the function currently does not handle mortalities); and the number of marked fish in the population prior to the ith sample (\code{M}).
+#'    \item \code{sum} A data frame that contains the the number of captured fish in the ith sample (\code{n}), the number of recaptured (i.e. previously marked) fish in the ith sample (\code{m}), the number of marked fish returned to the population following the ith sample (\code{R}; this will equal \code{n} as the function currently does not handle mortalities); the number of marked fish in the population prior to the ith sample (\code{M}); the number of fish first seen in the ith sample (\code{u}); the number of fish lasst seen in the ith sample (\code{v}); and the number of fish seen i times (\code{f}).
 #'    \item \code{methodB.top} A matrix that contains the top of the Method B table used for the Jolly-Seber method (i.e., a contingency table of capture sample (columns) and last seen sample (rows)).
 #'    \item \code{methodB.bot} A data.frame that contains the bottom of the Method B table used for the Jolly-Seber method (i.e., the number of marked fish in the sample (\code{m}), the number of unmarked fish in the sample (\code{u}), the total number of fish in the sample (\code{n}), and the number of marked fish returned to the population following the sample (\code{R}).
 #'    \item \code{m.array} A matrix that contains the the so-called \dQuote{m-array}.  The first column contains the number of fish captured on the ith event.  The columns labelled with \dQuote{cX} prefix show the number of fish originally captured in the ith row that were captured in the Xth event.  The last column shows the number of fish originally captured in the ith row that were never recaptured.
 #'  }
 #'
-#' @seealso See \code{\link{capHistConvert}} for a descriptions of capture history data file formats and how to convert between them.  See \code{\link{mrClosed}} and \code{\link{mrOpen}} for how to estimate abundance from the summarized capture history information.
+#' @seealso See \code{description} in \pkg{Rcapture} for \code{m.array} and some of the same values in \code{sum}.  See \code{\link{capHistConvert}} for a descriptions of capture history data file formats and how to convert between them.  See \code{\link{mrClosed}} and \code{\link{mrOpen}} for how to estimate abundance from the summarized capture history information.
 #'
 #' @section fishR vignette: \url{https://sites.google.com/site/fishrfiles/gnrl/MRClosed.pdf}, \url{https://sites.google.com/site/fishrfiles/gnrl/MROpen.pdf}
 #'
+#' @references Baillargeon, S. and Rivest, L.-P. (2007). Rcapture: Loglinear models for capture-recapture in R. Journal of Statistical Software, 19(5):1-31.
+#' 
 #' @keywords manip
+#' 
+#' @aliases capHistSum age.CapHist
 #' 
 #' @examples
 #' # data.frame with IDs in the first column
 #' data(PikeNYPartial1)
 #' head(PikeNYPartial1)
-#' capHistSum(PikeNYPartial1,cols2use=-1)
+#' ( ch1 <- capHistSum(PikeNYPartial1,cols2use=-1) )
+#' 
+#' # diagnostic plots
+#' plot(ch1)
+#' plot(ch1,what="f")
+#' plot(ch1,what="u")
 #'
 #' # An examle with only two sample events (for demonstration only)
-#' capHistSum(PikeNYPartial1,cols2use=-c(1,4:5))
+#' ( ch2 <- capHistSum(PikeNYPartial1,cols2use=-c(1,4:5)) )
 #'
+#' @rdname capHistSum
 #' @export
 capHistSum <- function(df,cols2use=NULL) {
   
@@ -52,15 +67,13 @@ capHistSum <- function(df,cols2use=NULL) {
   k <- dim(ch)[2]
   # Capture history frequencies
   res1 <- iSumCH(ch)
-  # Frequency of marked and unmarked fish in each sample
-  res2 <- iSumCHnum(ch,k)
   if (k==2) {
     d <- list(caphist=res1,sum=iSumCHnumPetersen(res1))
   } else {
     # Construct schnabel summary from marked & unmarked
-    res2.schnabel <- iSumCHnumSchnabel(res2,k)
+    res2.schnabel <- iSumCHnumSchnabel(ch,k)
     # Construct jolly bottom method B table from marked & unmarked
-    res2.jolly <- iSumCHnumJolly(res2,k)
+    res2.jolly <- iMethodBBot(ch,k)
     # Construct jolly top method B table
     res3 <- iMethodBTop(ch,k)
     # Construct m-array table from top of methodB table
@@ -124,27 +137,92 @@ iSumCHnumPetersen <- function(cht) {
 ##############################################################
 ## INTERNAL - Modify mark, recap, catch data to a Schnabel summary
 ##############################################################
-iSumCHnumSchnabel <- function(df,k) {
+iSumCHnumSchnabel <- function(ch,k) {
+  # Frequency of marked and unmarked fish in each sample
+  df <- iSumCHnum(ch,k)
   # Unmarked fish returned to the population after sample
   R <- df$n
   #   0 by default for last sample
   R[k] <- 0
   # Marked fish prior to sample
   M <- c(0,cumsum(df$n-df$m)[1:(k-1)])
+  # Number of fish seen for first time at i
+  u <- iGetUi(ch)
+  # Number of fish seen for lat time at i
+  v <- iGetVi(ch)
+  # Number of fish seen i times
+  f <- iGetFi(ch)
   # Return results as a data.frame
-  data.frame(n=df$n,m=df$m,R,M)
+  data.frame(n=df$n,m=df$m,R,M,u,v,f)
 }
+
+
+iGetFi <- function(chdf) {
+  # as.vector to strip attributes
+  as.vector(table(factor(rowSums(chdf),levels=1:ncol(chdf))) )
+}
+
+iGetVi <- function(chdf) {
+  events <- ncol(chdf)
+  vi <- numeric(events)
+  for (i in 1:events) {
+    if (i<(events-1)) {
+      # captured in the ith event, get ch for all events after that
+      tmp1 <- chdf[chdf[,i]==1,(i+1):events]
+      # If rowSums==0 then it was not captured again
+      tmpv <- table(rowSums(tmp1))
+      vi[i] <- ifelse(any(names(tmpv)=="0"),tmpv["0"],0)
+    } else if (i==(events-1)) {
+      # handle second to last event which produces a vector (can't use rowSums)
+      tmp1 <- chdf[chdf[,i]==1,events]
+      vi[i] <- length(tmp1)
+    } else {
+      # handle last event ... vi is just ni
+      vi[i] <- sum(chdf[,i])
+    }
+  }
+  vi
+}
+
+iGetUi <- function(chdf) {
+  events <- ncol(chdf)
+  ui <- numeric(events)
+  for (i in 1:events) {
+    if (i==1) {
+      # handle first event, ui=ni
+      ui[i] <- sum(chdf[,i])
+    } else if (i==2) {
+      tmp1 <- chdf[chdf[,i]==1,1:(i-1)]
+      ui[i] <- length(tmp1)
+    } else {
+      # captured in ith event, get ch for all events before that
+      tmp1 <- chdf[chdf[,i]==1,1:(i-1)]
+      # if rowSums==0 then it was not captured previously
+      tmpu <- table(rowSums(tmp1))
+      ui[i] <- ifelse(any(names(tmpu)=="0"),tmpu["0"],0)
+    }
+  }
+  ui
+}
+
 
 ##############################################################
 ## INTERNAL - Modify mark, recap, catch data to a Jolly (bottom Method B) summary
 ##############################################################
-iSumCHnumJolly <- function(df,k) {
+iMethodBBot <- function(ch,k) {
+  # Frequency of marked and unmarked fish in each sample
+  df <- iSumCHnum(ch,k)
   # Unmarked fish returned to the population after sample
   R <- df$n
   #   0 by default for last sample
   R[k] <- 0
-  # Essentially a re-ordering
-  t(data.frame(m=df$m,u=df$u,n=df$n,R=R))
+  # Put together a matrix (first 3 are re-ordering of df)
+  tmp <- rbind(df$m,df$u,df$n,R)
+  # rename rows and columns
+  rownames(tmp) <- c("m","u","n","R")
+  colnames(tmp) <- paste0("i=",1:ncol(tmp))
+  # return matrix
+  tmp
 }
 
 ##############################################################
@@ -170,6 +248,10 @@ iMethodBTop <- function(ch,k) {
       if (nrow(ch.tmp.i)==0) break
     } # end j loop
   } # end i loop
+  # relabel rows and columns
+  colnames(mb.top) <- paste0("i=",1:ncol(mb.top))
+  rownames(mb.top) <- paste0("j=",1:nrow(mb.top))
+  # return matrix
   mb.top
 }
 
@@ -179,7 +261,7 @@ iMethodBTop <- function(ch,k) {
 iMarray <- function(mb,smry) {
   events <- ncol(mb)
   # label rows and columns of mb (follow strategy in Rcapture)
-  rownames(mb) <- paste("i =",1:events)
+  rownames(mb) <- paste0("i=",1:events)
   colnames(mb) <- paste0("c",1:events)
   # put ni in first column
   mb[,1] <- smry$n
@@ -190,4 +272,33 @@ iMarray <- function(mb,smry) {
   colnames(mb)[events+1] <- "not recapt"
   # return matrix
   mb
+}
+
+#' @rdname capHistSum
+#' @export
+plot.CapHist <- function(x,what=c("u","f"),...) {
+  what <- match.arg(what,several.ok=TRUE)
+  tmp <- x$sum[,c("n","u","f")]
+  t <- nrow(tmp)
+  tmp$i <- 1:t
+  # scale the F variable
+  tmp$sf <- log(tmp$f/choose(t,tmp$i))
+  tmp$sf[which(tmp$f==0)] <- NA
+  # log of the U variable
+  tmp$su <- log(tmp$u)
+  tmp$su[which(tmp$u==0)] <- NA
+  # catch if two plots are being made
+  if (length(what)==2) {
+    old.par <- par(mfrow=c(1,2))
+    on.exit(par(old.par))
+  }
+  # make the fi plot
+  if ("f" %in% what) {
+    tmp2 <- tmp[!is.na(tmp$sf),]
+    plot(sf~i,data=tmp2,ylab="log(scaled fi)",xlab="Number of Times Captured",type="b",pch=16,...)
+  }
+  # make the ui plot
+  if ("u" %in% what) {
+    plot(su~i,data=tmp,ylab="log(ui)",xlab="Capture Event (i)",type="b",pch=16,...)
+  }
 }